+++ aiobungie/crate/__init__.py
@@ -73,7 +73,8 @@ __all__ = (
     "Artifact",
     "ArtifactPoint",
     "FetchableArtifact",
-    "ProfileCurrencies",
+    "ProfileItem",
+    "ProfileItemImpl",
 )
 
 from .activity import Activity
@@ -108,7 +109,8 @@ from .milestones import MilestoneItems
 from .profile import LinkedProfile
 from .profile import Profile
 from .profile import ProfileComponent
-from .profile import ProfileCurrencies
+from .profile import ProfileItem
+from .profile import ProfileItemImpl
 from .profile import ProfileProgression
 from .season import Artifact
 from .season import ArtifactPoint